clear interior + starting layouts
